Add Templar Variety
DAO templars are all the same.  Literally, most of the templars in DAO are clones of the same few characters (and despite the fact that female templar robes exist in the files and are perfectly usable, they're all men until Awakening).  This mod ports the medium templar armors from DAI, and distributs them to give the templars more variety.
URL: https://www.nexusmods.com/dragonage/mods/6003