Add No Outsiders in Orzammar
Orzammar is portrayed as pretty against outsiders and non-dwarves entering Orzammar, so it is wierd that several mercenaries of different races are found within the Carta and Proving matches. This mod changes these non-Dwarves to Dwarves. All of the mages were changed to 2-handed Berserkers. The Qunari Mercenary and Human Champion remain in their vanilla specializations.
URL: https://www.nexusmods.com/dragonage/mods/6240