Remove unused Crime Wave fix mod
The files in the Crime Wave Fix mod are duplicated in the Quinn's Fixpack, whose files take precedent following alpha sort rules. The functionality in the mod was never iucorporated into gameplay due to this conflict. In future it may be possible to incorporate some of the changes into the Quinn dialog, but that is not in scope of this commit.