Refactor repository to allow for both source and deployed top level folders
Initially the repository only contained deployed resources which did noot allow for any custom source objects to be stored. This commit moves the deployed resources to a new override folder and adds a new source folder that may be used to store files that should not exist in the DA game folders.