Move custom talk tables to mod tables
Custom talk tables are not read from the packages override folder, they need to be included as part of an addin. This commit combines the talk tables from the deployed mods into a new addin talk table for the mod.
Talk tables from the base game may be included in the packages folder which is why the More Detailed Tooltips mod does not require to be touched.