Fix strings in Spell Shaping modified
The strings for the Spell Shaping mod were missing from the mog string table. These strings have been both improved and added. The changes to the strings were to use the same tootip template as other skills and to also update the text to be more grammatically correct and consistent with other skills.