Add Camp Merchant Chest mod
Creates a storage chest by Bodahn Feddic's wagon in the Party Camp. Uses the Merchant UI, thereby providing categories, comparison with equipped gear, and far more space than a standard chest. The mod was merged into archid to allow it to be merged with autoloot from dains fixes to ensure that there is no attempt to autoloot it. As with all merged files objects are renamed to have an af_ prefix for easy identification.
URL: https://www.nexusmods.com/dragonage/mods/5979