Add The Rescue at Ishal mod
This mod extends the ending of the battle of Ostagar (mod contains spoilers). It adds two new cutscenes and two new pre-filmed movies, showing details of the devastating aftermath of the battle, what happened to Alistair and the Warden after the tower ambush. Lore-friendly. For all races, genders, and origins-NOT recommended for first-time players.
The cutscene files are too large for git so are added to the ignore file.
URL: https://www.nexusmods.com/dragonage/mods/4690