Store strings in Excel
Whilst it is possible to maintain the string tables in a text file having them in Excel simplifies things as the multi-line strings are able to be treated as a single entity when performing sorts or similar activities. This allows for simpler duplicate string ID checking which gets moer complicated the more mods are added.