Merge SpellShaping mod into archid
The spell shaping mod had a few minor bugs included in the script files so needed to be merged into the archid mod in order to recompile them.
As part of this process the changes to update dains fixes were also merged in, including both the event manager and effect manager.
The gda file entries from dains fixes that would have been overridden due to load order were removed to simplify the files.