Fix spell damage not being applied
Spells were not causing damage to targets due to my lack of understanding of how eventmanager overrides worked.
It appears that once an override is set for an event the base handler will not be called, even if ReleaseLock is called. In order for control to pass to the base handler it also needs to be configured in eventmanager.
This commit adds ability_core as the last override for the EVENT_TYPE_ABILITY_CAST_IMPACT event to allow for the standard processing if the spellshaping code does not handle it.