Allow changing logging level without recompilation
Thr current code has the logging level defined as a constant in the af_constants_h file. In order to change thsi level a recompilation is required. This is extremely restrictive and cumbersome. This commit changes the the location of the global log level definition to be in the DragonAge.ini file which is read when the module loads and stored in a module local varaible.
In order to do this the var_module.gda file needs to be updated directly to add the new variable. This is not ideal but it did not appear top be possible to extend the file using an m2da definition. The variable needs to be known by the toolkit at compilation time in order to be effective and the toolkit does not seem to respect the m2da.