Configure function level logging in DragonAge.ini file
The logging code has been updated to allow individual functions or groups of functions to dynamicaly set their maximum logging level in the ini file.
This allows for INFO, WARN & DEBUG logging statements to be added to the compiled code and switched on and off in groups rather than globally. An example would be that all script code for the unlock spell would use a logging group defined as UnlockSpell when calling the afLogxxxx function; then in the ini file UnlockSpell could be defined as having a max log level of Debug during mod development and then switched to Info once development is completed, or even switched off altogether.

To configure this in the DragonAge.ini file the following entries are recognised, with the config below having global logging set to DEBUG, but calls with the UnlockSpell log group defined not being written.
```
[Archid]
LogGlobalLevel=3
LogGroupsNone=UnlockSpell
LogGroupsInfo=
LogGroupsWarn=
LogGroupsDebug=
```

These entries are read when the module loads and can be optionally reread by running the following command in the console. This allows the values to be updated mid game
```
runscript af_zz_readini
```