Implement Dog Gift Tweaks mod
Reverse engineered the original mod to allow it to be created as an event manager listener rather than an override.
The original mod stated it provided +1% Melee Critical Chance, however this is not actually creature property. As an alternative this commit adds +2% Dodge Chance instead.
The archid logging system was added with a log group of DogGifts
URL: https://www.nexusmods.com/dragonage/mods/635