Merge Dains Equip Manager to archid mod
In order to simplify the code and allow for extension the equip manager has been merged into the archid mod. The denl_quickbar mod has been removed as this was always incompatable as it overrode the core event without providing the source coode so couldn't be included in the merge. The quickbar mod was a convenience, but not considered overly important as there are suitable alternatives.