Reduce dependancy on af_constants_h
As the af_constants_h file contains constants that are are functional specific it was being frequently updated, with the knock on effect of forcing recompilation of numerous scripts. In order to reduce the number of recompilations some of the constants that are functional specific have been moved into functionality specific scrips.