Merge Dains 'Attack resolution' tweaks
This mod component contains changes to core shared libraries than needs to be merged in to allow for all scripts to be compiled against the changed sources.
The mod description is below, taken from the Dains Fixes mod page:
* Attack resolution:
This is 5 mods in one, which are unfortunately inextricably linked. You can enable/disable each by editing the options_af.gda file with GDApp﻿.
    - Normalises the behaviour between auto-crit (autocrit effect e.g. pinpoint strike, death hex and attacking from stealth) and never-crit (dual strike, rapid shot), with each such effect canceling an opposing one out. Vanilla behaviour with both causes non-crit rolls to convert to crits and crit rolls to convert to hits, incentivising a low crit chance.
    ﻿- Allows backstabs to land when under the effects of Misdirection Hex, if the attack threatens a critical.
    ﻿- Implements the Messy Kills item property, doubling the chance of a deathblow.
    - Calculates attack timing with Dual Striking as the average of both weapons, rather than the main hand only.
    - Removes the evasion dodge animation that interrupts the rogue.
The various talent changes are all compiled with this tweak and will respect the options gda file, defaulting to vanilla behaviour if this tweak isn't included.