Add Dain's 'Suppressing Fire' fix
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Suppressing Fire:
Now only triggers on ranged attacks or talents, rather than any form of damage. E.g. tainted blade won't trigger it on yourself, and throwing an acid flask won't trigger it on enemies.