Add Dain's 'Lacerate' fix
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Lacerate:
Currently does absolutely nothing (the game checks if the character has a modal active, not a passive ability). Fixed it to apply the intended DoT.