Merge Dain's 'Modal exploration irritations' component
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Modal exploration irritations:
Removed the damage heartbeat from Blood Thirst, The Tainted Blade and Aura of Pain during exploration, and Berserk preventing stamina regen while exploring.