Merge Dain's 'Blizzard disabling effects' tweak
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Blizzard disabling effects:
Existing behaviour only checks once each to slip and freeze - when entering the blizzard, and on the third (globally, not per unit) heartbeat respectively. Blizzard now has a 12.5% chance to do each of those actions on every heartbeat (75% chance of no effect). The freeze may also now allow enemies to be shattered.