Merge Dain's 'Blood Control damage on resist' extra
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Blood Control damage on resist:
Now only applies damage when the target can't be charmed (due to having dialogue) or it passes the mental resistance check.