Merge Dain's 'Increases Hostility and Intimidation' fix
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:
* Increases Hostility and Intimidation:
Implemented this item effect in line with source comments. Now increases threat by 5 per attack (like Increases Hostility) and adds an effective 10 strength when evaluating an Intimidate roll.