Merge Dain's 'Don't dispel 3rd party modals' tweak
The component contains modified core libraries and is merged to allow all code to use the same code.
The description of the component from the Dain's Fixes mod site is:

* Don't dispel 3rd party modals:
Dispel no longer removes buffs from third party sustains. For example, it won't remove Haste provided by another mage. If dispel is cast on the mage sustaining the effect, it will still be deactivated.