Merge Dain's 'Autoloot kills' (conservative version) component
The autoiloot header file was already included in the code base, but was not hooked up to the event system.
The description of the component from the Dain's Fixes mod site is:
* Autoloot kills (conservative):
Automatically loots any enemy that does not hold a plot item or one that triggers an event on pickup. Flashes a warning if the party's inventory is full.