Merge Dain's 'Display received items' component
Added an option to enable/disable the functionality via the options file.
The description of the component from the Dain's Fixes mod site is:
Displays the name of any item(s) received. The colour and duration can be customised by editing the itemrcvd_options_af.gda file with GDApp. By default it's green and appears for 2 seconds.