Fix character respec component
There were several issues with the character respec component where the dialog was not working correctly and some files were not named well.
This commit attempts to resolve all these issues but has only been tested in the kings camp at ostagar so tests in further areas need to be performed.